Move console initialization into `console-logging-setup.js`
